must faster algorithm for rendering masked sprites

+++ sprites.h
@@ -38,7 +38,7 @@ public:
 
 	void drawBitmap(int16_t x, int16_t y,
 	  const uint8_t *bitmap, const uint8_t *mask,
-	  int16_t w, int16_t h, uint8_t draw_mode);
+	  int8_t w, int8_t h, uint8_t draw_mode);
 
 private:
